Edge_AI/0821/generic/requirements.txt
wheel==0.45.1
# numpy should be 2.3.0 for Ubuntu 24.04
#numpy==2.3.0
# numpy should be 2.2.6 for Ubuntu 22.04
numpy==2.2.6
pillow==11.2.1
matplotlib==3.10.3
scikit-learn==1.7.0
seaborn==0.13.2
